+++ src/main/java/org/realityforge/gwt/eventsource/client/event/MessageEvent.java
@@ -2,6 +2,7 @@ package org.realityforge.gwt.eventsource.client.event;
 
 import com.google.gwt.event.shared.EventHandler;
 import javax.annotation.Nonnull;
+import javax.annotation.Nullable;
 import org.realityforge.gwt.eventsource.client.EventSource;
 import org.realityforge.gwt.eventsource.client.event.MessageEvent.Handler;
 
@@ -21,12 +22,34 @@ public class MessageEvent
     return TYPE;
   }
 
+  @Nullable
+  private final String _lastEventId;
+  @Nonnull
+  private final String _messageType;
+  @Nonnull
   private final String _data;
 
-  public MessageEvent( @Nonnull final EventSource eventSource, @Nonnull final String data )
+  public MessageEvent( @Nonnull final EventSource eventSource,
+                       @Nullable final String lastEventId,
+                       @Nonnull final String messageType,
+                       @Nonnull final String data )
   {
     super( eventSource );
+    _lastEventId = lastEventId;
+    _messageType = messageType;
     _data = data;
+  }
+
+  @Nullable
+  public String getLastEventId()
+  {
+    return _lastEventId;
+  }
+
+  @Nonnull
+  public String getMessageType()
+  {
+    return _messageType;
   }
 
   @Nonnull